Don't run kernel check on openvz
This check is moot on openvz, as the host node controls the kernel.

I've also seen it create some weird errors such as (verbatim):
A newer kernel is installed, however the system has not been rebooted. running kernel: 2.6.32-042stab078.28, most recent installed kernel:

The "newer" kernel is blank as on openvz there may not be any kernel packages installed.